Refactor curriculum management to improve ownership verification and streamline module and lesson operations

+++ src/main/java/br/edu/ifsp/tcc/application/usecase/DeleteUserUseCase.java
@@ -1,7 +1,9 @@
 package br.edu.ifsp.tcc.application.usecase;
 
 import br.edu.ifsp.tcc.application.service.UserService;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
+import org.springframework.web.server.ResponseStatusException;
 
 @Component
 public class DeleteUserUseCase {
@@ -12,7 +14,10 @@ public class DeleteUserUseCase {
         this.userService = userService;
     }
 
-    public void execute(Long id) {
+    public void execute(Long id, Long authenticatedUserId) {
+        if (!id.equals(authenticatedUserId)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Acesso negado.");
+        }
         userService.deleteById(id);
     }
 }
